Adds favorites feature using local storage
Implements a new feature to allow users to save favorite meals locally.

Refactors the favorites repository and local storage to handle full meal objects instead of just meal IDs, storing the data as JSON strings in shared preferences.

Introduces a new `FavoritesBloc` to manage the state of the favorites page, displaying the list of favorite meals.

Adds a toggle to switch between list and grid views.